+++ hittable.h
@@ -10,7 +10,7 @@ struct hit_record {
   point3 p;
   vec3 normal;
   material *m;
-  double t;
+  float t;
   bool front_face;
 
   void set_face_normal(const ray &r, const vec3 &outward_normal);
@@ -19,7 +19,7 @@ struct hit_record {
 class hittable {
 public:
   virtual ~hittable() {}
-  virtual bool hit(const ray &r, double t_min, double t_max,
+  virtual bool hit(const ray &r, float t_min, float t_max,
                    hit_record &rec) const = 0;
 };
 #endif // HITTABLE_H_
