+++ hittable_list.cc
@@ -9,13 +9,13 @@ hittable_list::~hittable_list() {
   }
 }
 
-bool hittable_list::hit(const ray &r, double t_min, double t_max,
+bool hittable_list::hit(const ray &r, float t_min, float t_max,
                         hit_record &rec) const {
 
 
   hit_record temp_rec;
   bool hit_anything = false;
-  double closest_so_far = t_max;
+  float closest_so_far = t_max;
 
   for (auto o : objects_) {
     if (o->hit(r, t_min, t_max, temp_rec)) {
