+++ hittable_list.h
@@ -14,7 +14,7 @@ public:
     objects_.emplace_back(h);
   }
 
-  bool hit(const ray &r, double t_min, double t_max, hit_record &rec) const;
+  bool hit(const ray &r, float t_min, float t_max, hit_record &rec) const;
 
 private:
   std::vector<hittable *> objects_;
