+++ metal.h
@@ -8,13 +8,13 @@
 
 class metal : public material {
 public:
-  metal(const color &a, double f);
+  metal(const color &a, float f);
 
   bool scatter(const ray &r_in, const hit_record &rec, color &attenuation,
                ray &scattered) const override;
 
 private:
   color albedo_;
-  double fuzziness_;
+  float fuzziness_;
 };
 #endif // METAL_H_
