+++ ray.h
@@ -13,7 +13,7 @@ public:
   const point3 &origin() const { return origin_; }
   const vec3 &direction() const { return direction_; }
 
-  point3 at(double t) const { return origin_ + t * direction_; }
+  point3 at(float t) const { return origin_ + t * direction_; }
 
 private:
   point3 origin_;
