+++ vec3.h
@@ -8,20 +8,20 @@
 class vec3 {
   public:
     vec3() : v{0, 0, 0} {}
-    vec3(double x, double y, double z) : v{x, y, z} {}
+    vec3(float x, float y, float z) : v{x, y, z} {}
     vec3(const vec3 &other) {
       v[0] = other.v[0];
       v[1] = other.v[1];
       v[2] = other.v[2];
     }
 
-    double x() const { return v[0]; }
-    double y() const { return v[1]; }
-    double z() const { return v[2]; }
+    float x() const { return v[0]; }
+    float y() const { return v[1]; }
+    float z() const { return v[2]; }
 
     vec3 operator-() const { return vec3(-v[0], -v[1], -v[2]); }
-    double operator[](int i) const { return v[i]; }
-    double &operator[](int i) { return v[i]; }
+    float operator[](int i) const { return v[i]; }
+    float &operator[](int i) { return v[i]; }
 
     vec3 &operator+=(const vec3 &o) {
       v[0] += o.v[0];
@@ -30,22 +30,22 @@ class vec3 {
       return *this;
     }
 
-    vec3 &operator*=(double t) {
+    vec3 &operator*=(float t) {
       v[0] *= t;
       v[1] *= t;
       v[2] *= t;
       return *this;
     }
 
-    vec3 &operator/=(double t) { return *this *= 1 / t; }
+    vec3 &operator/=(float t) { return *this *= 1 / t; }
 
-    double length_squared() const {
+    float length_squared() const {
       return v[0] * v[0] + v[1] * v[1] + v[2] * v[2];
     }
-    double length() const { return std::sqrt(length_squared()); }
+    float length() const { return std::sqrt(length_squared()); }
 
     static vec3 random();
-    static vec3 random(double min, double max);
+    static vec3 random(float min, float max);
 
     static vec3 random_in_unit_sphere();
     static vec3 random_in_unit_disk();
@@ -53,10 +53,10 @@ class vec3 {
     static vec3 random_in_hemisphere(const vec3 &normal);
 
     static vec3 reflect(const vec3 &v, const vec3 &n);
-    static vec3 refract(const vec3 &r_in, const vec3 &n, double etai_over_etat);
+    static vec3 refract(const vec3 &r_in, const vec3 &n, float etai_over_etat);
 
   private:
-    double v[3];
+    float v[3];
 };
 
 using point3 = vec3;
@@ -72,12 +72,12 @@ vec3 operator-(const vec3 &u, const vec3 &v);
 
 vec3 operator*(const vec3 &u, const vec3 &v);
 
-vec3 operator*(const vec3 &v, double t);
-vec3 operator*(double t, const vec3 &v);
+vec3 operator*(const vec3 &v, float t);
+vec3 operator*(float t, const vec3 &v);
 
-vec3 operator/(vec3 v, double t);
+vec3 operator/(vec3 v, float t);
 
-double dot(const vec3 &u, const vec3 &v);
+float dot(const vec3 &u, const vec3 &v);
 
 vec3 cross(const vec3 &u, const vec3 &v);
